Balance start and finish recording calls
Keep a balanced count of startRecording vs finishRecording calls for
audio recording clients to address the desync between the audio policy
service and app ops that can occur when mic access is restricted.

The effect of this change should be that finishRecording is called
exactly when it was before, but *also* called if needed to balance the
successful startRecording calls, when no active recordings remain with
matching finish arguments.

This does not provide perfect state synchronization, as the current
system design impedes us from reliably determining app ops' own counter
value; hence, we wait until we know all active recordings are done
before sending our finish notifications. We might send more than needed,
but because this does not happen until all usage is complete, it should
not be a problem.

Test: atest CtsMediaAudioPermissionTestCases
Test: Manual: Mic privacy indicator still disappears when expected after
starting with mic access off and then canceling (one or more times) and
turning mic access back on, or unblocking, and ending recording.
Bug: b/190517602
Bug: b/293603271
Bug: b/363915467
Bug: b/374800336
Bug: b/374869085
Bug: b/377407253
Issue: calyxos#3123
Change-Id: I1a67234d8809898d47ab1845687f249b32831546